requirements.txt
numpy==2.1.3
pandas==2.3.3
scikit-learn==1.7.2
lightgbm==4.6.0
matplotlib==3.10.8
seaborn==0.13.2
joblib==1.5.3
malariagen_data==15.6.0
pysam==0.23.3
biopython==1.86
